added unit tests

+++ src/avatars/avatars.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
-import { Model, Types } from 'mongoose';
+import { Model } from 'mongoose';
 import { AvatarType } from './dto/avatar';
 import { Avatar } from './schemas/avatar.schema';
 import { readFileSync, unlink } from 'node:fs';
@@ -17,33 +17,34 @@ export class AvatarsService {
     avatar,
   }: {
     user: UserType;
-    avatar: Record<string, string | number>;
+    avatar: Express.Multer.File;
   }): Promise<unknown> {
     const buffer = readFileSync(avatar.path);
     const base64String = Buffer.from(buffer).toString('base64');
     const final = `data:${avatar.mimetype};base64,${base64String}`;
-    console.log(final, buffer);
     const avatarData: AvatarType = {
       userId: user._id,
       hash: final,
       file: avatar.path as string,
+      _id: '',
     };
     const newAvatar = await this.avatarModel.create(avatarData);
     return newAvatar;
   }
 
-  async findOne(id: string): Promise<Avatar> {
+  async findOne(id: string): Promise<Avatar | null> {
     return await this.avatarModel.findOne({ userId: id }).exec();
   }
 
   async delete(id: string) {
     const deletedAvatar = await this.avatarModel
-      .findOneAndDelete({ userId: new Types.ObjectId(id) })
+      .findOneAndDelete({ userId: id })
       .exec();
-    unlink(deletedAvatar.file, (err) => {
-      if (err) throw err;
-      console.log(`successfully deleted ${deletedAvatar.file}`);
-    });
+    if (deletedAvatar)
+      unlink(deletedAvatar.file, (err) => {
+        if (err) throw err;
+        console.log(`successfully deleted ${deletedAvatar.file}`);
+      });
     return deletedAvatar;
   }
 }
